Rename rules module to detections for better semantic clarity
- Rename rules.py -> detections.py
- Rename test_rules.py -> test_detections.py
- Update import references in __init__.py and test files
- All tests continue to pass
