+++ src/mcp_panther/panther_mcp_core/tools/detections.py
